requirements.txt
gmusicapi==7.0.0
spotipy==2.3.7
cached-property==1.3.0

appdirs==1.4.0
beautifulsoup4==4.4.1
cffi==1.5.0
cryptography==1.2.2
decorator==4.0.6
enum34==1.1.2
funcsigs==0.4
gpsoauth==0.0.4
httplib2==0.9.2
idna==2.0
ipaddress==1.0.16
MechanicalSoup==0.4.0
mock==1.3.0
mutagen==1.31
ndg-httpsclient==0.4.0
oauth2client==1.5.2
pbr==1.8.1
proboscis==1.2.6.0
protobuf==2.6.1
pyasn1==0.1.9
pyasn1-modules==0.0.8
pycparser==2.14
pycrypto==2.6.1
pyOpenSSL==0.15.1
python-dateutil==2.4.2
requests==2.9.1
rsa==3.3
six==1.10.0
spotipy==2.3.7
validictory==1.0.1
